Update UI Tests
